Large update: Please Read
Restarting with a different design.

1. Saving all the buffers to disk causes a lot of manual work. Going
   forward, we'll keep them in place in memory until we want to write
   out, and when we write out, we'll only do it with relevant data.

2. There is now a Global Object Manager that holds object data. The
   purpose of this change is that we can copy data that is otherwise
   lost, such as buffers with certain flags, before the copy to the GPU,
   and work out when it was referenced later. While this change does add
   the complexity of having to track many more references, it is an
   improvement overall.

3. Serialising the data should be a lot more streamlined now. When a
   frame is written to disk, we'll write it in a call-centric way.
   Chimera will need an update to match this change, and that project
   will be upgraded to convert to mesh-centricity during it's runtime.

Goals:
The number one goal of this project was, and still remains, to export
the customisation data of games with character customisation without
capturing the currently playing in-game animation.

While there are many challenges that remain to be overcome for this goal
to become a reality, I have a few ideas of possible ways to extract the
needed information from the shader pipeline.